requirements.txt
addict==2.4.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.40.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==3.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmengine==0.7.4
mmsegmentation==1.0.0
model-index==0.1.11
nest-asyncio==1.5.6
numpy==1.24.3
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prettytable==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
rich==13.4.2
scipy==1.10.1
scikit-image==0.22.0
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
termcolor==2.3.0
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
yapf==0.33.0
zipp==3.15.0
